pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-alternance"
version = "1.1.0"
description = "Agent IA qui automatise la recherche d'alternance : sourcing, scoring, candidatures, suivi"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Sidiné Coulibaly", email = "sidinecoulibalypro@gmail.com"},
]

dependencies = [
    # Scraping
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    # IA
    "anthropic>=0.25",
    # Documents
    "python-docx>=1.1",
    "fpdf2>=2.7",
    # Config
    "pyyaml>=6.0",
    # API Web
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",    # Pour tester l'API FastAPI
]

[tool.setuptools.packages.find]
include = [
    "sourcing*",
    "qualification*",
    "candidature*",
    "suivi*",
    "api*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
